reorganize channelz package to be more discoverable
pluralize EntryPerPage
